feat : 물리 매니저에 스레드 적용
중력 관련 버그픽스 포함

+++ Simple2DGameEngine/DCollisionData.cpp
@@ -36,9 +36,9 @@ float DCollisionData::CalculateMass()
 
 void DCollisionData::UpdatePosition(float deltaTime) {
     if(!freezeX)
-        localPosition.x += static_cast<int>(velocity.x * deltaTime);
+        localPosition.x += velocity.x * deltaTime;
     if (!freezeY)
-        localPosition.y += static_cast<int>(velocity.y * deltaTime);
+        localPosition.y += velocity.y * deltaTime;
     
     // 속도 위치 업데이트
     aabb.min = localPosition - (size * 0.5f);
